Fix copying secret to clipboard when secret is hidden
Copying doesn't work if the secret text is hidden, so we need to show it,
copy it, and hide it again. (This happens too fast to be visible/rendered.)

To help facilitate this, changed how toggle buttons are implemented.
Instead of two buttons that use different functions that are alternately
shown, use a single button with a toggle function, that changes the button
when it is toggled as appropriate. The toggle state is kept in a variable.