Require user to submit a form before actually fetching the secret
Avoids modifying state (i.e. deleting the secret) with a GET request, which
is a no-no. Caused issues when a service the link was posted to (e.g. a
chat service) loaded the link to try and generate a preview before the
receiving user could load it and get the secret.